% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shapefile.R
\name{create_shapefile}
\alias{create_shapefile}
\title{Create State- or Nation-Wide District Shapefile}
\usage{
create_shapefile()
}
\description{
\code{create_shapefile()} creates your designated shapefile
}
