% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lea_get.R
\name{lea_get}
\alias{lea_get}
\title{Get Nation-Wide District Shapefile}
\usage{
lea_get(path = NULL)
}
\arguments{
\item{path}{A character vector specifying a file path, such as:
path = "./test". Will attempt to create the directory if it does not
currently exist.}
}
\value{
An RDS object and shapefiles.
}
\description{
\code{lea_get()}downloads the U.S. district shapefile from GitHub.
}
