% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run {leaidr} Mapbox Shiny App Example}
\usage{
runExample()
}
\value{
A Shiny app.
}
\description{
\code{runExample()} runs a Shiny app example. A great tutorial on how to include Shiny apps in packages is here: https://deanattali.com/2015/04/21/r-package-shiny-app/
}
