% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lea_get.R
\name{lea_get}
\alias{lea_get}
\title{Get Nation-Wide District Shapefile}
\usage{
lea_get(state = NULL)
}
\arguments{
\item{state}{Either a FIPS code or two-digit abbreviation for the state to
obtain district shape files. If \code{NULL}, shape files for the entire US
will be downloaded. Can be a vector of multiple states.}
}
\value{
An RDS object and shapefiles.
}
\description{
\code{lea_get()}downloads the U.S. district shapefile from GitHub.
}
