% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaidr-package.R
\docType{package}
\name{leaidr-package}
\alias{leaidr}
\alias{leaidr-package}
\title{leaidr: U.S. School District Shapefiles}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Easier mapping of U.S. school districts by downloading and prepping NCES geography files.
}
\author{
\strong{Maintainer}: Isabella Velásquez \email{ivelasq@gmail.com}

Authors:
\itemize{
  \item Joshua Rosenberg \email{jmichaelrosenberg@gmail.com}
}

}
\keyword{internal}
